plugins: extract base class for Lexicon-based DNS authenticators
Introduce abstract classes to provide base functionality for Lexicon-
based DNS Authenticator plugins and corresponding test cases.

Refactor CloudXNS and DNSimple to use these new base classes.

Testing Done:
 * `tox -e py27`
 * `tox -e lint`
 * Manual Testing:
    * Used `certbot certonly --dns-dnsimple -d`, specifying an API token
      as a command line argument. Verified that a certificate was
      successfully obtained without user interaction.
    * Negative testing:
      * Invalid API token
      * Domain name not registered in DNSimple
 * `./tox.cover.sh` (1 relevant uncovered line: 50 of dns_dnsimple.py)
